Compile on Linux again.
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@4174 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackDummyDriver.cpp
@@ -56,7 +56,7 @@ int JackDummyDriver::Open(jack_nframes_t buffer_size,
         fEngineControl->fPeriod = 0;
         fEngineControl->fComputation = 500 * 1000;
         fEngineControl->fConstraint = 500 * 1000;
-        int buffer_size = (fWaitTime * fEngineControl->fSampleRate) / 1000000.0f;
+        int buffer_size = int((fWaitTime * fEngineControl->fSampleRate) / 1000000.0f);
         if (buffer_size > BUFFER_SIZE_MAX) {
             buffer_size = BUFFER_SIZE_MAX;
             jack_error("Buffer size set to %d ", BUFFER_SIZE_MAX);
